src/test/java/com/cebbus/bot/api/analysis/strategy/GlobalExtremaStrategyTest.java
package com.cebbus.bot.api.analysis.strategy;

import com.cebbus.bot.api.analysis.DataGenerator;
import org.jgap.Configuration;
import org.jgap.Gene;
import org.jgap.InvalidConfigurationException;
import org.jgap.impl.DoubleGene;
import org.jgap.impl.IntegerGene;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.ta4j.core.BarSeries;

import java.util.LinkedHashMap;
import java.util.Map;

import static org.junit.jupiter.api.Assertions.assertArrayEquals;
import static org.junit.jupiter.api.Assertions.assertEquals;

class GlobalExtremaStrategyTest {

    private BarSeries series;
    private CebStrategy strategy;
    private Number[] parameters;

    @BeforeEach
    void setUp() {
        this.series = DataGenerator.generateSeries();
        this.strategy = new GlobalExtremaStrategy(series);
        this.parameters = new Number[]{7, 0.996D, 1.004D};
    }

    @Test
    void getSeries() {
        assertEquals(this.series, this.strategy.getSeries());
    }

    @Test
    void getParameters() {
        assertArrayEquals(this.parameters, this.strategy.getParameters());
    }

    @Test
    void rebuild() {
        Number[] expected = new Number[]{1, 2D, 3D};
        this.strategy.rebuild(expected);

        assertEquals(expected, this.strategy.getParameters());
    }

    @Test
    void getParameterMap() {
        Map<String, Number> map = new LinkedHashMap<>(this.parameters.length);
        map.put("Price Bar Count", this.parameters[0]);
        map.put("High Coefficient", this.parameters[1]);
        map.put("Low Coefficient", this.parameters[2]);

        assertEquals(map, this.strategy.getParameterMap());
    }

    @Test
    void createGene() throws InvalidConfigurationException {
        Configuration conf = new Configuration();
        IntegerGene priceBarCount = new IntegerGene(conf, 1, 30);
        DoubleGene highCoefficient = new DoubleGene(conf, 0.990, 0.999);
        DoubleGene lowCoefficient = new DoubleGene(conf, 1.001, 1.009);

        assertArrayEquals(new Gene[]{priceBarCount, highCoefficient, lowCoefficient}, this.strategy.createGene(conf));
    }

}